magit-log-maybe-show-commit: fix log-follow case
When only the `log-follow' case is enabled, but not `log-online', then
the other window should only be updated to show the revision at point,
if it already shows another revision.  It should not to do merely
because that buffer exists (but is not visible in the selected frame).

To address this a new argument FRAME has to be added to
`magit-mode-get-buffer'.  If it turns out we can address the above issue
differently, then this argument might be removed again.